Re-enable `open_basedir`.
Most php libraries should not need access to the filesystem (beyond the
local project filesystem), so enabling `open_basedir` helps guarantee
that these libraries don't do anything fishy.  Overriding this default
is simple enough for any packages that need more permissions.

This also sets up PATH to include the local vendor/bin and global
$COMPOSER_HOME/vendor/bin paths.  These paths supercede the standard
PATH so that they can be relied on.

The composer script itself is installed to `$COMPOSER_HOME/vendor/bin`
(i.e., the `composer global` bin dir).  I don't foresee this being an
issue, but it is possible that it could cause conflicts with composer
packages.  If so, perhaps `$COMPOSER_HOME/bin` would be better.

The default cacert.pem that curl distributes is downloaded and used for
composer.  The composer installer typically does this itself, but its
logic conflicts a lot with `open_basedir`.  While it would technically
succeed in installing to `$COMPOSER_HOME/cacert.pem`, it would only do
so after trying to access restricted global certificate locations,
creating a lot of warnings.

I split off the `open_basedir` settings into a `basedir.ini` file in
order to make it simple to overwrite the defaults without mucking with
composer's dependencies.