Separate timezone into a separate file.
Some of PHP's date and time libraries throw warnings when the timezone
isn't set.  It's therefore a pretty good idea to always set it.

However, it should likely be set to a "correct" value, so making it easy
to configure is preferred.  By splitting it out into a separate file, it
makes it easy for child Dockerfile's to overwrite it.