Remove open_basedir configuration.
New versions of composer are wanting access to /etc/ for finding certs
and such which just makes things more difficult to work with.  I'd like
to still use open_basedir because I feel like it is important, but it
would be a lot of effort and would remain fragile to other future
changes.  This is worth revisiting in the future, however.