Remove separate build user and run as root.
Instead of bringing in the difficulty of a build user that needs to be
setup on the host, instead user namespaces can be used to provide a
similar level of security for the host system.  This should make this
image easier to work with, especially for custom additions and running
on non-Linux development machines.

Group setup is still necessary for volume mounting permissions to work
under many setups, but it's configurable by the user instead of
hard-coded by this docker image.

This should also hopefully improve the experience when trying to run
this image on an OSX or Windows host, especially using the newly in beta
project released by Docker:
https://blog.docker.com/2016/03/docker-for-mac-windows-beta/.