Switch from Arch Linux to Alpine Linux for smaller images.
Official docker images are beginning to migrate to Alpine Linux instead
of Arch Linux as it is much better suited for Docker's usecases.  The
small image sizes and easy configuration make it a preferred
distribution for a base.  PHP doesn't get as big of a benefit as other
environments because of the desire to bundle the PHP source into the
image for easy extension installation, but it does still reduce this
image from ~600 MB to ~450 MB which helps.

This switch to the official PHP image also makes it easier to use this
image - utilities for installing/enabling PHP extensions are already
included and it increases familiarity for users that have used the
official image previously.